Develop

+++ services/bank-simulator/src/main/java/com/sep/banksimulator/dto/qr/GenerateQrRequest.java
@@ -2,6 +2,8 @@ package com.sep.banksimulator.dto.qr;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -10,24 +12,15 @@ import lombok.*;
 public class GenerateQrRequest {
 
     private Long bankPaymentId;
-
     private Long pspPaymentId;
-
-    private Double amount;
-
+    private BigDecimal amount;
     private String currency;
-
     private String receiverAccount;
-
     private String receiverName;
-
     private String purpose;
-
     private String paymentCode;
-
     private String referenceNumber;
-
     private String stan;
-
     private String pspTimestamp;
+
 }
